refactor(app,wf-completion): inline tests from separate tests.rs files
